backend/requirements.txt
# ==================================
# CORE FRAMEWORK
# ==================================
Django==5.0.1
djangorestframework==3.14.0
django-cors-headers==4.3.1

# ==================================
# DATABASE
# ==================================
psycopg2-binary==2.9.10

# ==================================
# AUTHENTICATION & SECURITY
# ==================================
djangorestframework-simplejwt[blacklist]==5.3.1
PyJWT==2.8.0
django-ratelimit==4.1.0
argon2-cffi==23.1.0
django-axes==6.5.2

# ==================================
# PAYMENTS
# ==================================
mercadopago==2.2.1

# ==================================
# ASYNC TASKS
# ==================================
celery==5.3.4
redis==5.0.1
django-celery-beat==2.7.0
django-celery-results==2.5.1

# ==================================
# UTILITIES
# ==================================
python-decouple==3.8
python-dotenv==1.0.0
Pillow==10.4.0
qrcode==7.4.2

# ==================================
# EMAIL & NOTIFICATIONS
# ==================================
django-templated-mail==1.1.1

# ==================================
# PDF GENERATION
# ==================================
reportlab==4.0.7
weasyprint==60.1

# ==================================
# DEVELOPMENT TOOLS
# ==================================
django-extensions==3.2.3
ipython==8.18.1

# ==================================
# TESTING
# ==================================
pytest==7.4.3
pytest-django==4.7.0
factory-boy==3.3.0
faker==21.0.0

# ==================================
# CODE QUALITY
# ==================================
black==23.12.1
flake8==6.1.0
isort==5.13.2
bandit==1.7.5
safety==3.2.5

# ==================================
# PRODUCTION
# ==================================
gunicorn==21.2.0
whitenoise==6.6.0
dj-database-url==2.1.0
boto3==1.34.0  # Para S3 (certificados y archivos)

# ==================================
# API DOCUMENTATION
# ==================================
drf-yasg==1.21.7

